fix: 补充 np.frombuffer 缺失的 dtype 参数，修复缓冲区字节对齐错误
在 reshape 前未指定 dtype=np.uint8，导致默认 float64 要求 8 字节对齐，边长奇数时崩溃。

+++ core.py
@@ -139,7 +139,9 @@ def encode(ui: "UI") -> None:
 
     ui.set_loading_info("运行进度：正在映射图像。")
     result: Image.Image = Image.fromarray(
-        np.frombuffer(data).reshape(
+        np.frombuffer(
+            data, dtype=np.uint8
+        ).reshape(
             side_length, side_length, 4
             ),
         mode="RGBA"
